refactor: reuse functional wrappr for index and plugin
