+++ templates/components/plugin.js
@@ -1,36 +1,7 @@
 import Vue from 'vue'
+import { wrapFunctional } from './index'
+
 <% const components = options.getComponents() %>
-
-// nuxt/nuxt.js#8607
-function wrapFunctional (options) {
-  if (!options.functional) {
-    return options
-  }
-
-  const propKeys = Array.isArray(options.props) ? options.props : Object.keys(options.props || {})
-
-  return {
-    render(h) {
-      const attrs = {}
-      const props = {}
-
-      for (const key in this.$attrs) {
-        if (propKeys.includes(key)) {
-          props[key] = this.$attrs[key]
-        } else {
-          attrs[key] = this.$attrs[key]
-        }
-      }
-
-      return h(options, {
-        on: this.$listeners,
-        attrs,
-        props,
-        scopedSlots: this.$scopedSlots,
-      }, this.$slots.default)
-    }
-  }
-}
 
 const components = {
 <%= components.map(c => {
